backend/requirements.txt
fastapi
uvicorn
pydantic-ai
python-dotenv
pdfplumber
